This commit adds a script to fix compilation issues on macos
It adds a directory "mac" and a script "make_darwins.sh", following the pd "mac" directory model,
but it has very little to do with it.

After compiling and installing the library, this script places the darwin bundles
(one for each source) inside the ~/Documents/Pd/externals directory. It also
copies the 'model' dir and the generated pix_opencv-help.pd file, which are
not installed via makefile.